+++ Engine/gapi/vulkan/vallocator.h
@@ -65,7 +65,7 @@ class VAllocator {
 
     void getMemoryRequirements   (MemRequirements& out, VkBuffer buf);
     void getImgMemoryRequirements(MemRequirements& out, VkImage  img);
-    Allocation allocMemory(const MemRequirements& rq, const size_t heapId, const size_t typeId);
+    Allocation allocMemory(const MemRequirements& rq, const uint32_t heapId, const uint32_t typeId);
 
     bool commit(VkDeviceMemory dev, std::mutex& mmapSync, VkBuffer dest, size_t offset,
                 const void *mem, size_t count, size_t size, size_t alignedSz);
